ci: add routes for github pages
Previously, the application would not load correctly when deployed to github pages, it just loaded a
blank page. This had something to do with React Routes, which I do not understand at all because
they haven't been covered in the course yet. I managed to copy/paste some code from a medium article
to get it working. Here it is.